Anadiendo un Policy para seguridad

+++ app/Http/Controllers/VacanteController.php
@@ -103,6 +103,8 @@ class VacanteController extends Controller
      */
     public function edit(Vacante $vacante)
     {
+        $this->authorize('view', $vacante);
+
         $categorias = Categoria::all();
         $experiencias = Experiencia::all();
         $ubicaciones = Ubicacion::all();
@@ -125,6 +127,8 @@ class VacanteController extends Controller
      */
     public function update(Request $request, Vacante $vacante)
     {
+        $this->authorize('update', $vacante);
+
         // validación
         $data = $request->validate([
             'titulo' => 'required|min:6',
@@ -159,6 +163,8 @@ class VacanteController extends Controller
      */
     public function destroy(Vacante $vacante)
     {
+        $this->authorize('delete', $vacante);
+
         $vacante->delete();
         return response()->json([
             'ok' => true,
